Add STAC input and output to preprocessing.py
preprocessing.py now conforms to EO Application Package best practice
by interpreting its input not as a direct path to a data file but as
the path to a directory containing a STAC catalog containing an item
containing an asset giving the path to the actual input. Similarly,
after processing, preprocessing.py also writes a STAC catalog
describing the output data.